stream seeking bug.
All internal uses of std::function now supply an allocator; updated persistent identity to compile with custom memory management; fixed warning in SQS integration tests

S3 High Level Event Logging API

  Pretty simply Queue() and Flush() interface
  A simple test exists.

  The LogEvent class exists to give a strong concept of time to the event
  stream.

  Still to do:
  1. S3 bucket needs to be managed in the test, OR we change to
  mocks(thoughts here?)
  2. The actual format of the log is incredibly simple. There no log levels
  or anything, just a timestamp and a string. Do we want to expand this?

CMake-controlled curl client option for Windows: set CURL_DIR to a valid curl install directory

Rval ref specialization of BuildFunction added

Events are now converted to use Aws::Allocator while being tracked within the SDK

Compile fix for credential expiration

Liberalized timing on credential unit test

Make BuildFunction conditionally compiled since gcc doesn't support custom allocators to std::function

Fixed identity-management to use cognito-identity timestamp as double.

std::function allocator macro update; wait for queues to disappear before returning from delete

Synced brets content-type and better cmake stuff.

Merged version config conflict.

Fixes for MAC OSX and GNU builds.

Android build fixes

Test fixes; nested comment warning suppression due to auto-generated comments

Added export stuffs.Please enter the commit message for your changes. Lines starting

+++ aws-cpp-sdk-identity-management/include/aws/identity-management/auth/CognitoCachingCredentialsProvider.h
@@ -46,14 +46,14 @@ namespace Aws
 
             virtual CognitoIdentity::Model::GetCredentialsForIdentityOutcome GetCredentialsFromCognito() const = 0;
 
-            bool IsTimeExpired(long long expiry);
+            bool IsTimeExpired(double expiry);
 
             std::shared_ptr<CognitoIdentity::CognitoIdentityClient> m_cognitoIdentityClient;
             std::shared_ptr<PersistentCognitoIdentityProvider> m_identityRepository;
 
         private:
             AWSCredentials m_cachedCredentials;
-            std::atomic<long long> m_expiry;
+            std::atomic<double> m_expiry;
             std::mutex m_credsMutex;
         };
